pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[project]
authors = [{name = "Niccolò Cantù", email = "niccolo.cantu@nina.no"}]
dependencies = [
  "django>=5.2",
  "python-slugify>=8.0.1",
  "Pillow==11.1.0",
  "argon2-cffi",
  "psycopg[binary,pool]>=3.2.4",
  "backoff",
  "requests",
  # django-utils
  "django-health-check",
  "django-probes",
  "django-environ",
  "django-model-utils",
  "django-appconf>=1.0.6",
  "django-cors-headers>=4.3.1",
  "django-lifecycle",
  "django-postgres-extra",
  "django-polymorphic>=3.1.0",
  "rules>=3.3",
  "django-filter>=24.2",
  "django-widget-tweaks",
  # django-utils-app
  "django-solo",
  "django-treebeard>=4.7",
  "django-countries-plus",
  "django-languages-plus",
  # auth
  "django-allauth==65.5.0",
  "django-allauth-ui==1.7.1",
  # ui
  "django-crispy-forms",
  "fontawesomefree",
  "slippers",
  "crispy-tailwind",
  "django-tailwind",
  "django-tables2",
  "django-leaflet",
  # rest
  "django-rest-framework>=0.1.0",
  "drf-spectacular>=0.27.2",
  "drf-nested-routers",
  "dj-rest-auth",
  "drf-standardized-errors[openapi]",
  "djangorestframework-simplejwt",
  # gis
  "pygeometa @ git+https://github.com/nicokant/pygeometa@gbif-eml#egg=pygeometa",
  # "gdal==3.10.1",
  "pyproj",
  # queue
  "procrastinate[django]>=3.0.0",
  "setuptools>=75.8.0",
  "django-storages[s3]>=1.14.5",
  "django-taggit>=6.1.0",
  "django-autocomplete-light>=3.11.0",
  "django-view-breadcrumbs>=2.5.1",
  "django-auditlog>=3.0.0",
  "django-import-export>=4.3.7",
  "sentry-sdk>=1.40.5",
  "django-autoslug>=1.9.9",
  "pyjwt[crypto]>=2.9.0",
  "django-jsonform>=2.23.2",
  "django-svelte-jsoneditor>=0.4.2",
  "jsonfield>=3.1.0",
  "django-cotton>=2.0.3",
  "django-htmx>=1.23.0",
  "beautifulsoup4>=4.13.3",
  "lxml>=5.3.1",
  "xmltodict>=0.14.2",
  "fsspec>=2025.3.2",
  "aiohttp>=3.11.16",
  "pyarrow>=20.0.0",
  "s3fs>=2025.3.2",
  "pydantic>=2.11.4",
  "duckdb>=1.2.2",
  "fiona[s3]>=1.10.1",
  "rasterio>=1.4.3"
]
description = "Add your description here"
license = {text = "GPLv3"}
name = "dms"
readme = "README.md"
requires-python = ">=3.12"
version = "0.1.0"

[project.optional-dependencies]
dev = [
  "Werkzeug[watchdog]",
  "ipdb",
  "pre-commit",
  "django-debug-toolbar",
  "django-extensions",
  "ruff"
]
ldap = ["python-ldap", "django-auth-ldap"]
prod = ["gunicorn"]
test = [
  "pytest",
  "pytest-sugar",
  "pytest-django",
  "coverage",
  "factory-boy",
  "django-coverage-plugin"
]

[project.scripts]
dms = "dms.manage:main"
"manage.py" = "dms.manage:main"

# ==== Coverage ====
[tool.coverage.run]
include = ["metadata_dmsue/**"]
omit = ["*/migrations/*", "*/tests/*"]
plugins = ["django_coverage_plugin"]

[tool.django-stubs]
django_settings_module = "config.settings.test"

# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
# TODO: remove T002 when fixed https://github.com/Riverside-Healthcare/djLint/issues/687
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2

[tool.hatch]

[tool.hatch.metadata]
allow-direct-references = true

# ==== mypy ====
[tool.mypy]
check_untyped_defs = true
ignore_missing_imports = true
plugins = ["mypy_django_plugin.main"]
python_version = "3.11"
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
ignore_errors = true
# Django migrations should not produce any errors:
module = "*.migrations.*"

[tool.pylint]

[tool.pylint.DESIGN]
max-parents = 13

[tool.pylint.FORMAT]
max-line-length = 119

# ==== PyLint ====
[tool.pylint.MASTER]
django-settings-module = "config.settings.local"
load-plugins = ["pylint_django"]

[tool.pylint."MESSAGES CONTROL"]
disable = ["missing-docstring", "invalid-name"]

[tool.pylint.TYPECHECK]
generated-members = [
  "REQUEST",
  "acl_users",
  "aq_parent",
  "[a-zA-Z]+_set{1,2}",
  "save",
  "delete"
]

# ==== pytest ====
[tool.pytest.ini_options]
addopts = "--ds=dms.core.settings.test --reuse-db"
minversion = "6.0"
python_files = ["tests.py", "test_*.py"]

[tool.ruff]
fix = true

[tool.ruff.lint]
ignore = ["COM812"]
select = ["E", "W", "I", "F", "UP", "S", "B", "A", "COM", "LOG", "PTH", "Q"]

[tool.ruff.lint.per-file-ignores]
"**/migrations/*.py" = ["E501", "ANN"]
"**/tests/*.py" = ["S101"]

[tool.uv]
no-binary-package = ["fiona"]
